Added GameManager class that contains methods needed to get access to game data as leaderboard and game history. Methods return formatted data as DataTable. Its also simplify the code in forms.
