+++ Lab6/TicTacToeGame/TicTacToeDataAccess/TicTacToeDataAccess.csproj
@@ -62,6 +62,7 @@
   <ItemGroup>
     <Compile Include="DatabaseConnection.cs" />
     <Compile Include="DatabaseManager.cs" />
+    <Compile Include="GameManager.cs" />
     <Compile Include="PlayerManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
