+++ Lab6/TicTacToeGame/TicTacToeGame/MainWindow.xaml.cs
@@ -77,13 +77,13 @@ namespace TicTacToeGame
 
         private void OpenGameHistoryWindow(object sender, RoutedEventArgs e)
         {
-            GameHistoryWindow gameHistoryWindow = new GameHistoryWindow(dbManager);
+            GameHistoryWindow gameHistoryWindow = new GameHistoryWindow();
             gameHistoryWindow.Show();
         }
 
         private void OpenLeaderboardWindow(object sender, RoutedEventArgs e)
         {
-            LeaderboardWindow leaderboardWindow = new LeaderboardWindow(dbManager);
+            LeaderboardWindow leaderboardWindow = new LeaderboardWindow();
             leaderboardWindow.Show();
         }
     }
