+++ Lab6/TicTacToeGame/TicTacToeGame/View/LeaderboardWindow.xaml.cs
@@ -6,23 +6,18 @@ namespace TicTacToeGame
 {
     public partial class LeaderboardWindow : Window
     {
-        private DatabaseManager _databaseManager;
+        private GameManager _gameManager;
 
-        public LeaderboardWindow(DatabaseManager databaseManager)
+        public LeaderboardWindow()
         {
             InitializeComponent();
-            _databaseManager = databaseManager;
+            _gameManager = new GameManager();
             LoadLeaderboard();
         }
 
         private void LoadLeaderboard()
         {
-            var data = _databaseManager.ExecuteQuery(
-                "SELECT PlayerID, Username, COUNT(*) AS GamesWon " +
-                "FROM Players INNER JOIN Games ON Players.PlayerID = Games.WinnerID " +
-                "GROUP BY PlayerID, Username " +
-                "ORDER BY GamesWon DESC");
-            LeaderboardListView.ItemsSource = data.DefaultView;
+            LeaderboardListView.ItemsSource = _gameManager.GetLeaderboard().DefaultView;
         }
     }
 }
